added a performance tracking class

+++ src/main/scala/rca/auditor/LinkageSource.scala
@@ -5,8 +5,8 @@ import scala.beans.BeanProperty
 class LinkageSource (_name:String) extends Serializable {
 
 
-  @BeanProperty var population: Int = _
-  @BeanProperty var eligiblePopulation: Int = _
+  @BeanProperty var population: Long = _
+  @BeanProperty var eligiblePopulation: Long = _
   @BeanProperty val name:String = _name
 
 }
